# Update escaping of attribute and text HTML characters in SSR